% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_hclust.R
\name{fast_hclust_R}
\alias{fast_hclust_R}
\title{Fast Hierarchical Agglomerative Clustering (Pure R)}
\usage{
fast_hclust_R(dist_matrix, method = "single", dendrogram = TRUE)
}
\arguments{
\item{dist_matrix}{A square symmetric distance matrix}

\item{method}{Linkage method: "single", "complete", or "average"}

\item{dendrogram}{Logical, whether to plot the dendrogram}
}
\value{
An object of class 'hclust'
}
\description{
Fast Hierarchical Agglomerative Clustering (Pure R)
}
