% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_hclust.R
\name{fast_hclust_Rcpp}
\alias{fast_hclust_Rcpp}
\title{Fast HAC via Rcpp backend}
\usage{
fast_hclust_Rcpp(dist_matrix, method = "single", dendrogram = TRUE)
}
\arguments{
\item{dist_matrix}{Matrice de distance symétrique}

\item{method}{Méthode de linkage}

\item{dendrogram}{Afficher le dendrogramme}
}
\value{
Objet de classe hclust
}
\description{
Fast HAC via Rcpp backend
}
