% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dijkstra_graphGenerator.R
\name{generate_random_graph}
\alias{generate_random_graph}
\title{Generate a Random Weighted Graph}
\usage{
generate_random_graph(n, p)
}
\arguments{
\item{n}{Integer. The number of vertices in the graph.}

\item{p}{Numeric. The probability of an edge existing between two vertices.}
}
\value{
A data frame with three columns: 
 \itemize{
   \item{"v1"}: The first vertex of the edge.
   \item{"v2"}: The second vertex of the edge.
   \item{"cost"}: A random weight assigned to the edge (sampled from a normal distribution).
   }
}
\description{
This function generates a random weighted graph with `n` vertices and `m` edges,
where `m` is determined based on the probability `p` of an edge existing 
between any two vertices.
}
\examples{
generate_random_graph(10, 0.5)

}
