% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_1_insertion.R
\name{insertion_sort}
\alias{insertion_sort}
\title{Insertion Sort Algorithm}
\usage{
insertion_sort(v)
}
\arguments{
\item{v}{A numeric vector to be sorted.}
}
\value{
A sorted numeric vector with NA values excluded.
}
\description{
Sorts a numeric vector using the insertion sort algorithm, excluding NA values.
}
\examples{
insertion_sort(rnorm(100))
insertion_sort(sample(c(NA, 1:100)))
}
