% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_hclust.R
\name{naive_hclust_R}
\alias{naive_hclust_R}
\title{Naive HAC avec choix de la méthode de linkage}
\usage{
naive_hclust_R(dist_matrix, method = "single", dendrogramme = TRUE)
}
\arguments{
\item{dist_matrix}{Une matrice de distance symétrique n x n}

\item{method}{Méthode de linkage : "single", "complete" ou "average"}

\item{dendrogramme}{Booléen : TRUE pour afficher le dendrogramme}
}
\value{
An object of class 'hclust' containing merge history, heights, and order
}
\description{
Effectue un clustering hiérarchique agglomératif (HAC) avec une méthode naïve.
}
