% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_hclust.R
\name{naive_hclust_Rcpp}
\alias{naive_hclust_Rcpp}
\title{Naive HAC avec choix de la méthode de linkage (Rcpp)}
\usage{
naive_hclust_Rcpp(dist_matrix, method = "single", dendrogramme = TRUE)
}
\arguments{
\item{dist_matrix}{Une matrice de distance symétrique n x n}

\item{method}{Méthode de linkage : "single", "complete" ou "average"}

\item{dendrogramme}{Booléen : TRUE pour afficher le dendrogramme}
}
\value{
Un objet de type hclust
}
\description{
Effectue un clustering hiérarchique agglomératif (HAC) avec une méthode naïve.
}
