% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_Algorithms.R
\name{heap_sort}
\alias{heap_sort}
\title{Heap sort algorithm}
\usage{
heap_sort(v, type = "notRecursive")
}
\arguments{
\item{v}{an unsorted vector of numeric data}

\item{type}{there are two versions for the heap building : a recursive one and a direct one}
}
\value{
the sorted vector
}
\description{
Sorting by insertion with a heap structure
}
