% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heap_sort_Rcpp}
\alias{heap_sort_Rcpp}
\title{Heap sort algorithm using C++}
\usage{
heap_sort_Rcpp(v)
}
\arguments{
\item{v}{an unsorted vector of numeric data}
}
\value{
a sorted vector
}
\description{
Heap sort algorithm using C++
}
