% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorting_Algorithms.R
\name{insertion_sort}
\alias{insertion_sort}
\title{Insertion sort algorithm}
\usage{
insertion_sort(v)
}
\arguments{
\item{v}{an unsorted vector of numeric data}
}
\value{
the sorted vector
}
\description{
Sorting by insertion
}
