% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insertion_sort_Rcpp}
\alias{insertion_sort_Rcpp}
\title{Insertion sort algorithm using C++}
\usage{
insertion_sort_Rcpp(v)
}
\arguments{
\item{v}{an unsorted vector of numeric data}
}
\value{
a sorted vector
}
\description{
Insertion sort algorithm using C++
}
